Fixing tests after moving action from menu to maze
